I have no idea what I'm doing but let's try this anyways
I like the Lua programming language from an aesthetics point of view,
and I feel like learning Go and using what I've learned from Crafting
Interpreters, so I'm going to create a bytecode compiler and interpreter
for a language I'm calling `glua`, for Go and Lua. It's not going to be
as cool or fully featured as Lua but that's sort of my reference implementation.