Read file by UTF characters and print them out
This is a stepping stone to building a scanner that will produce
tokens. I think Go has built-in scanning functionality, but I want
to be able to split by any whitespace and by comments, and I may
even add multiline strings, which will require custom comment
scanning that I have a feeling can't be done by the std lib.