Implement subtraction, fix semicolon bug
Subtraction is easy enough because it's in the same precedence
as addition and it doesn't matter if you left to right or right to left.

I also removed the revert when the error is io.EOF because that caused
a panic. Plus I made a note about my confusion around how Lua
semicolons work in the grammar so I remember to go back to it