Add table accessor and attribute setter
There's a bunch of stuff packed into this change. It started with
a new test to get attributes from tables and one to set attributes.
As I started searching out how to add the syntax, I got a bit
stumped, because unlike the rest of the grammar, you don't know
whether a thing is going to be a getter or a setter until you've
parsed a whole expression to the left of an equal sign. You also
have to put the getters much higher on the precedence heirarchy.

So in order to decide how to move forward, I peeked back at how
the interpreter from Crafting Interpreters did this, and remembered
it used inheritance. So, I went back and changed all the functions
in the compiler to return the same Node interface, and added to
that interface a method that changes a getter to a setter, which
is only implemented on expressions that can be setters.

Other small changes include:

- Clearing errors in the VM on loop
- Removing interface types other than Node from the AST
- Removing all "wrapping" types unless they affect bytecode
- Fix printing to stdout instead of stderr in trace

I also changed some of the variable names to be longer and more
descriptive because it doesn't cost anything and makes the code
more readable for the methods.